refactor: /utils 하위 유틸함수 용도에 따라 분류
- getSomthing, renderSomthing 두가지 종류로 분류
- 두 분류에 해당하지 않는 함수가 추가되면 /utils/etc에 분류
- 복잡하지 않은 컴포넌트의 경우 컴포넌트 내부에 함수를 선언하여 사용하는것을 지향
