+++ src/components/common/modals/PrimaryModal.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import ReactModal from 'react-modal';
 import CheckButton from '../buttons/CheckButton';
 import { calcRatio } from '@lib/styles/theme';
-import renderCongratsIllust from '@lib/utils/common/renderCongratsIllust';
+import renderCongratsIllust from '@lib/utils/render/renderCongratsIllust';
 
 interface PrimaryModalProps {
   /**
