+++ src/components/home/pending/PendingDongilItem.tsx
@@ -2,7 +2,7 @@ import SuggestBadge from '@components/common/badges/SuggestBadge';
 import { modals } from '@components/common/modals/Modals';
 import useModals from '@lib/hooks/useModals';
 import { EDongilStatus } from '@lib/types/TDongilStatus';
-import { getDate } from '@lib/utils/common/getDate';
+import { getDate } from '@lib/utils/get/getDate';
 import { IDongil } from '@store/slices/walkingDongilsSlice';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
