+++ src/components/layout/AppBar.tsx
@@ -1,10 +1,10 @@
 import { useNavigate, useParams } from 'react-router-dom';
 import styled, { css } from 'styled-components';
 import { ReactComponent as Arrow } from '@assets/icons/arrow-left.svg';
-import getColorByLevel from '@lib/utils/common/getColorByLevel';
 import { isUndefined } from 'util';
 import { text } from 'stream/consumers';
 import { TLevel } from '@lib/types/TLevel';
+import getColorByLevel from '@lib/utils/get/getColorByLevel';
 
 interface AppBarProps {
   /**
