+++ src/lib/types/TLevel.ts
@@ -1,2 +1,2 @@
-// level 0은 홈 페이지 뱅키임당 빼꼼 배경을 위한것입니다.
+// level: -4은 홈 페이지 뱅키임당 빼꼼 배경을 위한것입니다.
 export type TLevel = 1 | 2 | 3 | 4 | -4 | 5;
