--- src/lib/types/TRiskLevel.ts
@@ -1,1 +1,0 @@
-export type TRiskLevel = '안정' | '보통' | '위험';
