+++ src/lib/utils/get/getContractEndDate.ts
@@ -1,6 +1,6 @@
 import moment from 'moment';
 
-export function getContractEndDate(createdAt: string, weeks: number) {
+function getContractEndDate(createdAt: string, weeks: number) {
   const createdDate = new Date(createdAt);
   const endDate = new Date(createdDate);
   const timeStamp = endDate.setDate(createdDate.getDate() + 7 * weeks - 1);
@@ -8,3 +8,5 @@ export function getContractEndDate(createdAt: string, weeks: number) {
   return dateObject;
   // return moment(endDate).format('YY.MM.DD');
 }
+
+export default getContractEndDate;
