+++ src/lib/utils/render/renderHomeBackground.tsx
