+++ src/lib/utils/render/renderItemIllustForCreate.tsx
